Add basic support for the Publons author widget.
The code searches the page for a #publonsWidget container and
extracts any Publons author id it can find within the container.
It then renders the Publons widget inplace.

The code requires jQuery and assumes that the VIVO site has aded
a "Publons widget" data property that people can populate. A
lighter-weight approach might be to look for a Publons url in
the websites section instead. A more robust solution might define
an ORNG gadget.